Allow defining OBC commands as mixins
