+++ integration_tests/obc/__init__.py
@@ -1,0 +1,7 @@
+from SerialPortTerminal import SerialPortTerminal
+from obc import OBC
+
+__all__ = [
+    'SerialPortTerminal',
+    'OBC'
+]
